+++ unlock-js/CHANGELOG.md
@@ -1,6 +1,18 @@
 # Changes
 
+## 0.2.5
+- Allow `walletService` to properly connect to `UnlockProvider`
+
+## 0.2.4
+
+- Increased gas limit to deploy Unlock contract
+
+## 0.2.3
+
+- Adding support for v11 of the smart contracts.
+
 ## 0.2.2
+
 - `unlock-provider` gains a property `isUnlock` that will allow Unlock-specific
   provider configuration.